Working on getting all of the utilities functioning.
Changed the bin directory to execs so that bin directories built from
setuptools could be ignored in the future.
Working on the docs, starting with README.rst.
Got the difference function working with a call to scale_several,
which required some changes. Ensured that all of the test-*
executables in the testing directories worked and had the right
output and comments. Moved Binner to powderx.py so that
cctbx code and non-cctbx code could be separated, which is a
necessity cuased by from __future__ import division in cctbx.